Removed unnecessary constructors in Models and Replaced salary datatype from Double to BigDecimal

+++ src/main/java/com/groovify/vinylshopapi/controllers/EmployeeController.java
@@ -8,6 +8,7 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import java.math.BigDecimal;
 import java.net.URI;
 import java.util.List;
 
@@ -27,8 +28,8 @@ public class EmployeeController {
             @RequestParam(required = false) String firstName,
             @RequestParam(required = false) String lastName,
             @RequestParam(required = false) String jobTitle,
-            @RequestParam(required = false) Double minSalary,
-            @RequestParam(required = false) Double maxSalary,
+            @RequestParam(required = false) BigDecimal minSalary,
+            @RequestParam(required = false) BigDecimal maxSalary,
             @RequestParam(required = false) String country,
             @RequestParam(required = false) String city,
             @RequestParam(defaultValue = "id") String sortBy,
